test(testing): sidebar и sync suite types вынесены из RSSReaderTests
Sidebar query, presentation и screen-state tests выделены в отдельные feature-oriented files, а CrossDeviceReadingScenarioTests переведён на standalone suite type
