+++ RSSReaderTests/RSSReaderTests.swift
@@ -7,252 +7,6 @@ import Testing
 @MainActor
 struct RSSReaderTests {
     @Test
-    func sourcesSidebarQuerySnapshotAggregatesUnreadAndStarredStateForFeeds() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let feeds = try harness.insertFeeds(
-            urls: [
-                "https://example.com/sidebar-feed-one.xml",
-                "https://example.com/sidebar-feed-two.xml"
-            ]
-        )
-        let firstFeed = try #require(feeds.first)
-        let secondFeed = try #require(feeds.last)
-
-        let unreadArticle = try harness.insertArticle(
-            feed: firstFeed,
-            externalID: "sidebar-unread",
-            url: "https://example.com/articles/unread",
-            title: "Unread Article"
-        )
-        let starredArticle = try harness.insertArticle(
-            feed: secondFeed,
-            externalID: "sidebar-starred",
-            url: "https://example.com/articles/starred",
-            title: "Starred Article"
-        )
-        _ = try harness.insertArticle(
-            feed: secondFeed,
-            externalID: "sidebar-read",
-            url: "https://example.com/articles/read",
-            title: "Read Article"
-        )
-
-        let stateService = try #require(harness.dependencies.articleStateService)
-        _ = try stateService.toggleStarred(article: starredArticle, at: .now)
-        _ = try stateService.markAsRead(feedID: secondFeed.id, articleExternalID: "sidebar-read", at: .now)
-        _ = unreadArticle
-
-        let snapshot = try harness.dependencies.sourcesSidebarQueryService?.fetchSnapshot()
-        let resolvedSnapshot = try #require(snapshot)
-
-        #expect(resolvedSnapshot.feeds.map(\.id) == feeds.map(\.id))
-        #expect(resolvedSnapshot.feeds.map(\.unreadCount) == [1, 1])
-        #expect(resolvedSnapshot.feeds.map(\.starredCount) == [0, 1])
-        #expect(resolvedSnapshot.unreadSmartCount == 2)
-        #expect(resolvedSnapshot.starredSmartCount == 1)
-        #expect(resolvedSnapshot.starredFeedIDs == [secondFeed.id])
-    }
-
-    @Test
-    func sidebarCountPresentationUsesUnreadCountersForAllItemsAndUnreadFilters() {
-        let feed = FeedSidebarItem(
-            feed: Feed(id: UUID(), url: "https://example.com/feed.xml", title: "Feed"),
-            unreadCount: 3,
-            starredCount: 2
-        )
-        let folder = FolderSidebarGroup(name: "Tech", feeds: [feed])
-
-        #expect(
-            SidebarCountPresentation.smartCount(
-                for: .allItems,
-                unreadSmartCount: 5,
-                starredSmartCount: 2
-            ) == 5
-        )
-        #expect(
-            SidebarCountPresentation.smartCount(
-                for: .unread,
-                unreadSmartCount: 5,
-                starredSmartCount: 2
-            ) == 5
-        )
-        #expect(SidebarCountPresentation.feedCount(for: feed, filter: .allItems) == 3)
-        #expect(SidebarCountPresentation.feedCount(for: feed, filter: .unread) == 3)
-        #expect(SidebarCountPresentation.folderCount(for: folder, filter: .allItems) == 3)
-        #expect(SidebarCountPresentation.folderCount(for: folder, filter: .unread) == 3)
-    }
-
-    @Test
-    func sidebarCountPresentationUsesStarredCountersForStarredFilter() {
-        let firstFeed = FeedSidebarItem(
-            feed: Feed(id: UUID(), url: "https://example.com/feed-one.xml", title: "Feed One"),
-            unreadCount: 4,
-            starredCount: 1
-        )
-        let secondFeed = FeedSidebarItem(
-            feed: Feed(id: UUID(), url: "https://example.com/feed-two.xml", title: "Feed Two"),
-            unreadCount: 2,
-            starredCount: 3
-        )
-        let folder = FolderSidebarGroup(name: "Tech", feeds: [firstFeed, secondFeed])
-
-        #expect(
-            SidebarCountPresentation.smartCount(
-                for: .starred,
-                unreadSmartCount: 6,
-                starredSmartCount: 4
-            ) == 4
-        )
-        #expect(SidebarCountPresentation.feedCount(for: firstFeed, filter: .starred) == 1)
-        #expect(SidebarCountPresentation.feedCount(for: secondFeed, filter: .starred) == 3)
-        #expect(SidebarCountPresentation.folderCount(for: folder, filter: .starred) == 4)
-    }
-
-    @Test
-    func sidebarSubtitleFormatterReturnsSyncingTitleForSyncingState() {
-        let formatter = SidebarSubtitleFormatter()
-
-        #expect(formatter.text(for: .syncing) == "Syncing...")
-    }
-
-    @Test
-    func sidebarSubtitleFormatterReturnsPlaceholderWhenNoRefreshDateIsAvailable() {
-        let formatter = SidebarSubtitleFormatter()
-
-        #expect(formatter.text(for: .idle(lastUpdatedAt: nil)) == "Not updated yet")
-    }
-
-    @Test
-    func sidebarSubtitleFormatterFormatsTodayRefreshDate() {
-        let formatter = SidebarSubtitleFormatter()
-        let calendar = Calendar.current
-        let now = Date()
-        let refreshDate = calendar.date(
-            bySettingHour: 9,
-            minute: 41,
-            second: 0,
-            of: now
-        ) ?? now
-
-        let expectedText = "Today at \(refreshDate.formatted(date: .omitted, time: .shortened))"
-
-        #expect(formatter.text(for: .idle(lastUpdatedAt: refreshDate)) == expectedText)
-    }
-
-    @Test
-    func sidebarSubtitleFormatterFormatsYesterdayRefreshDate() {
-        let formatter = SidebarSubtitleFormatter()
-        let calendar = Calendar.current
-        let now = Date()
-        let yesterday = calendar.date(byAdding: .day, value: -1, to: now) ?? now
-        let refreshDate = calendar.date(
-            bySettingHour: 21,
-            minute: 15,
-            second: 0,
-            of: yesterday
-        ) ?? yesterday
-
-        let expectedText = "Yesterday at \(refreshDate.formatted(date: .omitted, time: .shortened))"
-
-        #expect(formatter.text(for: .idle(lastUpdatedAt: refreshDate)) == expectedText)
-    }
-
-    @Test
-    func sidebarSubtitleFormatterFormatsOlderRefreshDateWithAbbreviatedDate() {
-        let formatter = SidebarSubtitleFormatter()
-        let refreshDate = Calendar.current.date(byAdding: .day, value: -7, to: Date()) ?? .distantPast
-
-        let expectedText = refreshDate.formatted(date: .abbreviated, time: .shortened)
-
-        #expect(formatter.text(for: .idle(lastUpdatedAt: refreshDate)) == expectedText)
-    }
-
-    @Test
-    func sidebarToolbarStateMarksSyncingStateAndUsesSyncingSubtitle() {
-        let state = SidebarToolbarState(refreshStatus: .syncing)
-
-        #expect(state.subtitle == "Syncing...")
-        #expect(state.isSyncing)
-    }
-
-    @Test
-    func sidebarToolbarStateMarksIdleStateAndUsesFormattedSubtitle() {
-        let refreshDate = Calendar.current.date(byAdding: .hour, value: -2, to: Date()) ?? Date()
-        let formatter = SidebarSubtitleFormatter()
-        let expectedSubtitle = formatter.text(for: .idle(lastUpdatedAt: refreshDate))
-        let state = SidebarToolbarState(refreshStatus: .idle(lastUpdatedAt: refreshDate))
-
-        #expect(state.subtitle == expectedSubtitle)
-        #expect(state.isSyncing == false)
-    }
-
-    @Test
-    func sidebarScreenStateExposesPrimaryLoadingStateThroughDerivedViewState() {
-        let state = SidebarScreenState()
-
-        let viewState = state.derivedViewState(
-            filter: .allItems,
-            expandedFolderNames: []
-        )
-
-        #expect(state.phase == .loading)
-        #expect(viewState.primaryLoadingState?.title == "Loading Sources")
-        #expect(viewState.placeholder == nil)
-        #expect(viewState.shouldDisableScrolling)
-        #expect(viewState.smartRows.isEmpty)
-    }
-
-    @Test
-    func sidebarScreenStateBuildsLoadedDerivedViewStateFromSnapshot() {
-        let feed = Feed(
-            id: UUID(),
-            url: "https://www.theverge.com/rss/index.xml",
-            title: "The Verge",
-            folder: Folder(name: "Tech")
-        )
-        let feedSidebarItem = FeedSidebarItem(
-            feed: feed,
-            unreadCount: 2,
-            starredCount: 1
-        )
-        let snapshot = SourcesSidebarSnapshotDTO(
-            feeds: [feedSidebarItem],
-            unreadSmartCount: 2,
-            starredSmartCount: 1,
-            starredFeedIDs: [feed.id]
-        )
-        let state = SidebarScreenState.previewLoaded(snapshot: snapshot)
-
-        let viewState = state.derivedViewState(
-            filter: .allItems,
-            expandedFolderNames: ["Tech"]
-        )
-
-        #expect(state.phase == .loaded)
-        #expect(viewState.primaryLoadingState == nil)
-        #expect(viewState.placeholder == nil)
-        #expect(viewState.smartRows.map(\.item) == [.allItems])
-        #expect(viewState.smartRows.first?.count == 2)
-        #expect(viewState.folderRows.count == 2)
-        #expect(viewState.ungroupedFeedRows.isEmpty)
-        #expect(viewState.shouldDisableScrolling == false)
-
-        guard case .folder(let folderRow)? = viewState.folderRows.first else {
-            Issue.record("Expected first folder row in SidebarScreenDerivedViewState")
-            return
-        }
-        #expect(folderRow.name == "Tech")
-        #expect(folderRow.isExpanded)
-
-        guard case .feed(let feedRow)? = viewState.folderRows.last else {
-            Issue.record("Expected nested feed row in SidebarScreenDerivedViewState")
-            return
-        }
-        #expect(feedRow.title == "The Verge")
-        #expect(feedRow.isIndented)
-    }
-
-    @Test
     func articleScreenStateStartsWithNoSelectionPlaceholder() {
         let state = ArticleScreenState()
         let viewState = state.derivedViewState()
